requirements.txt
streamlit>=1.35
langchain>=0.2
langchain-community>=0.2
langchain-google-genai>=2.0.0
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
pdfplumber>=0.11.4
pandas>=2.2
python-dotenv>=1.0.1
openai 
pymupdf 
python-dotenv
apify-client
langchain-openai
langchain_groq